refactor: rename to scope
